Support multiple language game data
- Created `xiv-gen-macros` crate with `FromCsv` derive macro.
- Refactored `xiv-gen` to use manually defined structs, reducing binary size and supporting multiple languages.
- Updated `xiv-gen-db` to generate compressed bincode files for EN, JA, DE, and FR.
- Updated `ultros` backend to serve language-specific game data.
- Updated `ultros-client` to fetch game data for the user's language.
- Fixed extensive compilation errors in `ultros-app` due to struct changes.
- Improved `SubrowKey` parsing and added missing fields for various game entities.

+++ ultros-frontend/ultros-app/src/routes/retainers.rs
@@ -30,7 +30,7 @@ impl From<(ItemId, bool)> for ItemSortKey {
             let sort_weight = sort_category
                 .get(&ItemSortCategoryId(item.item_sort_category as i32))
                 .map(|category| category.param)?;
-            Some(Self(sort_weight, item.key_id.0, hq))
+            Some(Self(sort_weight as u8, item.key_id.0, hq))
         };
         inner().unwrap_or(Self(u8::MAX, i32::MAX, hq))
     }
